Delete text UvR2021
Lions hunt by ambush. Fanning out, they form a semicircle, with the smaller lionesses herding the prey towards the center. A lion’s vision is roughly six times more sensitive to light than humans, giving them a distinct advantage when hunting at night. Lion’s claws are retractable, reaching up to 1 ½ inches in length, allowing for the tremendous control when it needs to go in for a kill. Lions can also reach speeds of up to 50 miles per hour in short bursts and can jump up to 36 feet.  So, it is easy to see why the lion is referred to as the “king of the jungle”, which leads us to our next fact...